requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.20.0
matplotlib>=3.4.0
Pillow>=8.2.0
tqdm>=4.60.0
opencv-python>=4.5.0
reedsolo>=1.5.4
tensorboard>=2.8.0
scikit-image>=0.18.0